+++ src/game/editor/editor.cpp
@@ -6951,7 +6951,7 @@ void CEditor::RenderEnvelopeEditor(CUIRect View)
 					{
 						m_PopupEnvelopeSelectedPoint = i;
 						static SPopupMenuId s_PopupCurvetypeId;
-						Ui()->DoPopupMenu(&s_PopupCurvetypeId, Ui()->MouseX(), Ui()->MouseY(), 80, NUM_CURVETYPES * 14.0f + 10.0f, this, PopupEnvelopeCurvetype);
+						Ui()->DoPopupMenu(&s_PopupCurvetypeId, Ui()->MouseX(), Ui()->MouseY(), 80, static_cast<float>(NUM_CURVETYPES) * 14.0f + 10.0f, this, PopupEnvelopeCurvetype);
 					}
 				}
 			}
